+++ src/core/hle/D3D8/Direct3D9/Direct3D9.cpp
@@ -61,6 +61,7 @@
 #include "common\input\InputManager.h"
 #include "common/util/strConverter.hpp" // for utf8_to_utf16
 #include "VertexShaderSource.h"
+#include "Timer.h"
 
 #include <imgui.h>
 #include <backends/imgui_impl_dx9.h>
@@ -2144,36 +2145,6 @@ static LRESULT WINAPI EmuMsgProc(HWND hWnd, UINT msg, WPARAM wParam, LPARAM lPar
     }
 
     return S_OK; // = Is not part of D3D8 handling.
-}
-
-// More precise sleep, but with increased CPU usage
-void SleepPrecise(std::chrono::steady_clock::time_point targetTime)
-{
-	using namespace std::chrono;
-	// If we don't need to wait, return right away
-
-	// TODO use waitable timers?
-	// TODO fetch the timer resolution to determine the sleep threshold?
-	// TODO adaptive wait? https://blat-blatnik.github.io/computerBear/making-accurate-sleep-function/
-
-	// Try to sleep for as much of the wait as we can
-	// to save CPU usage / power
-	// We expect sleep to overshoot, so give ourselves some extra time
-	// Note currently we ask Windows to give us 1ms timer resolution
-	constexpr auto sleepThreshold = 2ms; // Minimum remaining time before we attempt to use sleep
-
-	auto sleepFor = (targetTime - sleepThreshold) - steady_clock::now();
-	auto sleepMs = duration_cast<milliseconds>(sleepFor).count();
-
-	// Sleep if required
-	if (sleepMs >= 0) {
-		Sleep((DWORD)sleepMs);
-	}
-
-	// Spin wait
-	while (steady_clock::now() < targetTime) {
-		;
-	}
 }
 
 std::chrono::steady_clock::time_point GetNextVBlankTime()
